refactor: DRY improvments for genesis cli tets
