+++ x/oracle/client/cli/gen_pricefeeder_delegation_test.go
@@ -1,23 +1,14 @@
 package cli_test
 
 import (
-	"context"
 	"fmt"
 	"testing"
 
-	moduletestutil "github.com/cosmos/cosmos-sdk/types/module/testutil"
-
 	"github.com/NibiruChain/nibiru/app"
+	"github.com/NibiruChain/nibiru/x/common/testutil"
 	"github.com/NibiruChain/nibiru/x/oracle/client/cli"
 
-	"github.com/cometbft/cometbft/libs/log"
-	"github.com/cosmos/cosmos-sdk/client"
 	"github.com/cosmos/cosmos-sdk/client/flags"
-	"github.com/cosmos/cosmos-sdk/server"
-	"github.com/cosmos/cosmos-sdk/types/module"
-	"github.com/cosmos/cosmos-sdk/x/genutil"
-	genutiltest "github.com/cosmos/cosmos-sdk/x/genutil/client/testutil"
-	"github.com/spf13/viper"
 	"github.com/stretchr/testify/require"
 )
 
@@ -67,25 +58,8 @@ func TestAddGenesisPricefeederDelegation(t *testing.T) {
 	for _, tc := range tests {
 		tc := tc
 		t.Run(tc.name, func(t *testing.T) {
-			logger := log.NewNopLogger()
-
-			home := t.TempDir()
-			cfg, err := genutiltest.CreateDefaultTendermintConfig(home)
-			require.NoError(t, err)
-
-			testModuleBasicManager := module.NewBasicManager(genutil.AppModuleBasic{})
-			appCodec := moduletestutil.MakeTestEncodingConfig().Codec
-			err = genutiltest.ExecInitCmd(testModuleBasicManager, home, appCodec)
-			require.NoError(t, err)
-
-			serverCtx := server.NewContext(viper.New(), cfg, logger)
-			clientCtx := client.Context{}.WithCodec(appCodec).WithHomeDir(home)
-
-			ctx := context.Background()
-			ctx = context.WithValue(ctx, client.ClientContextKey, &clientCtx)
-			ctx = context.WithValue(ctx, server.ServerContextKey, serverCtx)
-
-			cmd := cli.AddGenesisPricefeederDelegationCmd(home)
+			ctx := testutil.SetupClientCtx(t)
+			cmd := cli.AddGenesisPricefeederDelegationCmd(t.TempDir())
 			cmd.SetArgs([]string{
 				fmt.Sprintf("--%s=%s", cli.FlagValidator, tc.validator),
 				fmt.Sprintf("--%s=%s", cli.FlagPricefeeder, tc.pricefeeder),
